+++ app/controllers/properties_controller.rb
@@ -42,6 +42,14 @@ class PropertiesController < ApplicationController
       if params[:property] && params[:property][:amenities_processed] == '1'
         params[:property][:amenities_processed] = Time.zone.today
       end
+      if params[:property] && params[:property][:lot_fees_paid]
+        if params[:property][:lot_fees_paid] == '1'
+          params[:property][:lot_fees_paid_on] = Time.zone.today
+        else
+          params[:property][:lot_fees_paid_on] = nil
+        end
+        params[:property].delete(:lot_fees_paid)
+      end
       if @property.update(property_params)
         resulting_location = property_url(@property)
         # if update is from index, return to index
@@ -76,6 +84,8 @@ class PropertiesController < ApplicationController
       params.require(:property).permit(
         :amenities_processed,
         :for_sale,
+        :lot_fees_paid,
+        :lot_fees_paid_on,
         :membership_eligible,
         :section,
         :street_number,