+++ app/decorators/property_decorator.rb
@@ -27,9 +27,7 @@ class PropertyDecorator < Draper::Decorator
   end
 
   def toggleable_amenities_processed?
-    h.form_with model: property, data: { controller: 'autosave'} do |f|
-      f.check_box :amenities_processed?, data: { action: 'autosave#save' }
-    end
+    toggleable_date_as_boolean(model: property, attribute_name: :amenities_processed)
   end
 
   def toggleable_for_sale?
@@ -46,4 +44,21 @@ class PropertyDecorator < Draper::Decorator
     #   onchange: "this.setAttribute('data-params', 'checked=' + this.checked*this.checked)",
     #   data: { remote: true, url: h.property_path(property), method: :patch }
   end
+
+  def toggleable_lot_fees_paid?
+    toggleable_date_as_boolean(model: property, attribute_name: :lot_fees_paid_on, boolean_attribute_name: 'lot_fees_paid?')
+  end
+
+  private
+
+  def toggleable_date_as_boolean(model:, attribute_name:, boolean_attribute_name: "#{attribute_name}?")
+    h.form_with model: model, data: { controller: 'autosave'} do |f|
+      data_attrs = { action: 'autosave#save' }
+      attribute_value = model.send(attribute_name)
+      data_attrs[:tooltip] = "On: #{attribute_value} (#{h.time_ago_in_words(attribute_value)} ago)" if attribute_value
+      f.check_box boolean_attribute_name, data: data_attrs
+    end
+  end
+
+
 end