+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_03_15_012603) do
+ActiveRecord::Schema[7.0].define(version: 2023_11_02_222201) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
   enable_extension "fuzzystrmatch"
@@ -130,6 +130,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_15_012603) do
     t.integer  "section"
     t.boolean  "for_sale"
     t.date     "amenities_processed"
+    t.date     "lot_fees_paid_on"
 
     t.index ["street_name"], :name=>"index_properties_on_street_name"
     t.index ["street_number"], :name=>"index_properties_on_street_number"